config/requirements.txt
flask
flask-bootstrap
redis
gitpython
